refactor(web): split landing page into section components [M6]
Extract 2077-line monolithic page.tsx into 16 focused section components
in web/src/components/landing/. Page.tsx is now ~120 lines that compose
sections with shared state passed via props.

Components: NavigationSection, HeroSection, DemoVideoSection, StatsSection,
AIFeaturesSection, HowItWorksSection, FeatureShowcasesSection, MidPageCTASection,
SolutionsSection, TestimonialsSection, SecuritySection, PricingSection,
FAQSection, FinalCTASection, StickyBottomBar, FooterSection

Shared utilities (Reveal, AnimatedStat, FAQItem, scrollTo) extracted to shared.tsx.
